Updating javadocs for tools classes; reorganizing files for more logical structue
