+++ src/main/java/com/coveros/selenified/selenium/Assert.java
@@ -23,8 +23,8 @@ package com.coveros.selenified.selenium;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.coveros.selenified.tools.OutputFile;
-import com.coveros.selenified.tools.OutputFile.Success;
+import com.coveros.selenified.OutputFile;
+import com.coveros.selenified.OutputFile.Success;
 
 /**
  * Assert will handle all verifications performed on the actual application
