+++ src/main/java/com/coveros/selenified/selenium/element/Contains.java
@@ -22,8 +22,8 @@ package com.coveros.selenified.selenium.element;
 
 import java.util.Arrays;
 
-import com.coveros.selenified.tools.OutputFile;
-import com.coveros.selenified.tools.OutputFile.Success;
+import com.coveros.selenified.OutputFile;
+import com.coveros.selenified.OutputFile.Success;
 
 /**
  * Contains extends Asserts to provide some additional verification
