+++ src/main/java/com/coveros/selenified/services/HTTP.java
@@ -29,7 +29,7 @@ import java.net.URL;
 import org.apache.commons.codec.binary.Base64;
 import org.testng.log4testng.Logger;
 
-import com.coveros.selenified.tools.General;
+import com.coveros.selenified.utilities.General;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
