+++ src/test/java/integration/ActionDoIT.java
@@ -13,7 +13,7 @@ import com.coveros.selenified.exceptions.InvalidBrowserException;
 import com.coveros.selenified.selenium.App;
 import com.coveros.selenified.selenium.Selenium.Browser;
 import com.coveros.selenified.selenium.Selenium.Locator;
-import com.coveros.selenified.tools.Selenified;
+import com.coveros.selenified.Selenified;
 
 public class ActionDoIT extends Selenified {
