+++ src/test/java/integration/AssertStateIT.java
@@ -6,7 +6,7 @@ import org.testng.annotations.Test;
 
 import com.coveros.selenified.selenium.App;
 import com.coveros.selenified.selenium.Selenium.Locator;
-import com.coveros.selenified.tools.Selenified;
+import com.coveros.selenified.Selenified;
 
 public class AssertStateIT extends Selenified {
