+++ src/test/java/sample/POMSampleIT.java
@@ -7,7 +7,7 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import com.coveros.selenified.selenium.App;
-import com.coveros.selenified.tools.Selenified;
+import com.coveros.selenified.Selenified;
 
 public class POMSampleIT extends Selenified {
