+++ src/test/java/unit/ExceptionTest.java
@@ -6,7 +6,7 @@ import org.testng.annotations.Test;
 
 import com.coveros.selenified.exceptions.InvalidBrowserException;
 import com.coveros.selenified.selenium.Selenium.Browser;
-import com.coveros.selenified.tools.TestSetup;
+import com.coveros.selenified.utilities.TestSetup;
 
 public class ExceptionTest {
