+++ src/test/java/unit/ResponseTest.java
@@ -10,9 +10,9 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import com.coveros.selenified.OutputFile;
 import com.coveros.selenified.selenium.Selenium.Browser;
 import com.coveros.selenified.services.Response;
-import com.coveros.selenified.tools.OutputFile;
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
 import com.google.gson.JsonArray;
