+++ src/test/java/unit/SeleniumTest.java
@@ -3,10 +3,10 @@ package unit;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import com.coveros.selenified.OutputFile.Success;
 import com.coveros.selenified.exceptions.InvalidBrowserException;
 import com.coveros.selenified.selenium.Selenium.Browser;
 import com.coveros.selenified.selenium.Selenium.DriverSetup;
-import com.coveros.selenified.tools.OutputFile.Success;
 
 public class SeleniumTest {
